account/requirements.txt
flask
flask-sqlalchemy
pymysql
requests
python-dotenv
flask-cors
